tests/com_detection/requirements.txt
pyserial==3.5
crccheck==1.3.0
pip==24.2
pycrc==0.9.2
pyvesc==1.0.5
setuptools==72.1.0
wheel==0.43.0
